Осталось написать функцию записи символов в CGRAM

+++ sources/Device/src/Modules/WH1602B/WH1602B.cpp
@@ -248,7 +248,7 @@ void WH1602B::lcdHigh(uint8 data)
 }
 
 
-void WH1602B::LoadSymbolToCGA(uint8, const uint8 [7])
+void WH1602B::LoadSymbolToCGA(int, const uint8 [7])
 {
 
 }
